+++ wtforms_alchemy/fields.py
@@ -477,7 +477,7 @@ class PhoneNumberField(StringField):
     .. _SQLAlchemy-Utils:
        https://github.com/kvesteri/sqlalchemy-utils
 
-    :param country_code:
+    :param region:
         Country code of the phone number.
     :param display_format:
         The format in which the phone number is displayed.
@@ -485,11 +485,16 @@ class PhoneNumberField(StringField):
     widget = TelInput()
     error_msg = u'Not a valid phone number value'
 
-    def __init__(self, label=None, validators=None, country_code='US',
-                 display_format='national',
-                 **kwargs):
+    def __init__(
+        self,
+        label=None,
+        validators=None,
+        region='US',
+        display_format='national',
+        **kwargs
+    ):
         super(PhoneNumberField, self).__init__(label, validators, **kwargs)
-        self.country_code = country_code
+        self.region = region
         self.display_format = display_format
 
     def _value(self):
@@ -514,7 +519,7 @@ class PhoneNumberField(StringField):
                 try:
                     self.data = PhoneNumber(
                         valuelist[0],
-                        self.country_code
+                        self.region
                     )
                     if not self.data.is_valid_number():
                         raise ValueError(self.gettext(self.error_msg))