+++ CHANGES.rst
@@ -4,6 +4,13 @@ Changelog
 Here you can see the full list of changes between each WTForms-Alchemy release.
 
 
+0.16.1 (2016-05-11)
+^^^^^^^^^^^^^^^^^^^
+
+- Updated SQLAlchemy-Utils requirement to 0.32.6
+- Fixed PhoneNumberType conversion (#102)
+
+
 0.16.0 (2016-04-20)
 ^^^^^^^^^^^^^^^^^^^